+++ src/routes/(app)/planning/life/+page.svelte
@@ -419,6 +419,7 @@
         </div>
         <div class="buttons">
           <a class="button is-light" href="/planning/life/whatif">What-If Scenarios</a>
+          <a class="button is-light" href="/planning/life/drawdown">Tax-Aware Drawdown</a>
           <a class="button is-light" href="/planning/life/goals">Manage Goals</a>
         </div>
       </div>